\name{erc}
\alias{erc}
\title{
Estimates remaining color
}
\description{
Estimates number of remaining color tags in the field
}
\usage{
erc(usedcombinations, yearusedcombinations, currentyear = NA,
  yearsurvival = NA, lifespan = NA, hide_color = NA)
}

\arguments{
  \item{usedcombinations}{
Pre used combinations
}
  \item{yearusedcombinations}{
The year in which the combination was used.
}
  \item{currentyear}{
Current year.
}
  \item{yearsurvival}{
An estimation of the proportion of animals that survive between years.
}
  \item{lifespan}{
Combinations older the lifespan will be automatically disregard.
}
  \item{hide_color}{
Color(s) to be hide in the estimation of remain colors tags present in nature.
}
}

\value{
A estimation of the number of remaining color tags in the field.
}

\author{
Biagolini-Jr.
}

\seealso{
scy
}
\examples{
## The function is currently defined as
data(pre_used) # Data example
erc(pre_used[,1:4],pre_used[,5],2019,0.85, hide_color="EMPTY")
}

