% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym_functions.R
\docType{data}
\name{pre_used}
\alias{pre_used}
\title{Pre-used combinations combinations}
\format{A data frame with 1200 observations on the following 5 variables.
\describe{
\item{\code{Tag_1}}{a factor with levels \code{Black} \code{Brown} \code{Dark_Blue} \code{EMPTY} \code{Gray} \code{Green} \code{Light_Blue} \code{Orange} \code{Pink} \code{Red} \code{White} \code{Yellow}}
\item{\code{Tag_2}}{a factor with levels \code{Black} \code{Brown} \code{Dark_Blue} \code{Gray} \code{Green} \code{Light_Blue} \code{Metal} \code{Orange} \code{Pink} \code{Red} \code{White} \code{Yellow}}
\item{\code{Tag_3}}{a factor with levels \code{Black} \code{Brown} \code{Dark_Blue} \code{EMPTY} \code{Gray} \code{Green} \code{Light_Blue} \code{Orange} \code{Pink} \code{Red} \code{White} \code{Yellow}}
\item{\code{Tag_4}}{a factor with levels \code{Black} \code{Brown} \code{Dark_Blue} \code{Gray} \code{Green} \code{Light_Blue} \code{Metal} \code{Orange} \code{Pink} \code{Red} \code{White} \code{Yellow}}
\item{\code{Year}}{a numeric vector}
}}
\source{
Simulated database
}
\usage{
data(pre_used)
}
\description{
Pre-used combinations combinations
}
\details{
Simulated database to example of registers of color tag  sequences usage. Columns 1, 2 ,3 and 4 represent the tag code, and column 5 is the date of tagging.
}
\keyword{datasets}
