\name{vfrequency}
\alias{vfrequency}
\title{
Variable frequency tag sequence sample
}
\description{
Create combinations with defined sample probability to each colors.
}
\usage{
vfrequency(ntag, colorsname, colorsf, nspecial = 0, name1 = "Metal", 
name2 = "EMPTY", location1 = 1, location2 = 2, nspecial1 = 1, nspecial2 = 1)
}
\arguments{
\item{ntag}{Number of tag to be used in each animal.}

\item{colorsname}{Names/Code of color tags to be sample.}

\item{colorsf}{Frequencies/ratio for color sample.}

\item{nspecial}{Number of special tags/codes, such as metallic, "EMPTY", or flag (min 0, max 2).}

\item{name1}{Name of special tag 1.}

\item{name2}{Name of special tag 2.}

\item{location1}{Position (or group of positions) to special band 1.}

\item{location2}{Position (or group of positions) to special band 2.}

\item{nspecial1}{Number of special tag 1 that will be present in all sequences genetated.}

\item{nspecial2}{Number of special tag 2 that will be present in all sequences genetated.}
}
\value{
A sequencie of tags
}

\author{
Biagolini-Jr
}
\seealso{
allequal
genseq
lifexp
}
\examples{
tcol<-c("Black","Blue","Brown","Gray","Green","Pink","Purple","Red","White","Yellow")
p<-c(1,2,5,1,2,2,4,5,8,5)
genseq(30, 4, tcol, gen_method="vfrequency", colorsf=p) 
}
