\name{pre_used}
\alias{pre_used}
\docType{data}
\title{
Pre-used combinations combinations
}
\description{
Simulated data of pre-used combinations
}
\usage{data("pre_used")}
\format{
  A data frame with 1200 observations on the following 5 variables.
  \describe{
    \item{\code{Tag_1}}{a factor with levels \code{Black} \code{Brown} \code{Dark_Blue} \code{EMPTY} \code{Gray} \code{Green} \code{Light_Blue} \code{Orange} \code{Pink} \code{Red} \code{White} \code{Yellow}}
    \item{\code{Tag_2}}{a factor with levels \code{Black} \code{Brown} \code{Dark_Blue} \code{Gray} \code{Green} \code{Light_Blue} \code{Metal} \code{Orange} \code{Pink} \code{Red} \code{White} \code{Yellow}}
    \item{\code{Tag_3}}{a factor with levels \code{Black} \code{Brown} \code{Dark_Blue} \code{EMPTY} \code{Gray} \code{Green} \code{Light_Blue} \code{Orange} \code{Pink} \code{Red} \code{White} \code{Yellow}}
    \item{\code{Tag_4}}{a factor with levels \code{Black} \code{Brown} \code{Dark_Blue} \code{Gray} \code{Green} \code{Light_Blue} \code{Metal} \code{Orange} \code{Pink} \code{Red} \code{White} \code{Yellow}}
    \item{\code{Year}}{a numeric vector}
  }
}
\details{
Simulated database to example of registers of color tag  sequences usage. Columns 1, 2 ,3 and 4 represent the tag code, and column 5 is the date of tagging.
}
\source{
Simulated database
}

\examples{
data(pre_used)
}
\keyword{datasets}
